crash only extension with minimal core classes modifications
