+++ src/com/sun/javatest/CrashOnlyParameters.java
@@ -1,0 +1,5 @@
+package com.sun.javatest;
+
+public interface CrashOnlyParameters extends Parameters {
+
+}
