+++ src/com/sun/javatest/DefaultTestRunner.java
@@ -27,18 +27,12 @@
 package com.sun.javatest;
 
 import com.sun.javatest.util.BackupPolicy;
-import com.sun.javatest.util.Crash;
 import com.sun.javatest.util.I18NResourceBundle;
 
 import java.io.PrintWriter;
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Traditional implementation of the test execution engine which has been
@@ -163,7 +157,7 @@ public class DefaultTestRunner extends TestRunner {
         }
     }
 
-    protected boolean runTest(TestDescription td) {
+    protected TestResult runTestGiveResult(TestDescription td){
         WorkDirectory workDir = getWorkDirectory();
         TestResult result = null;
 
@@ -206,12 +200,6 @@ public class DefaultTestRunner extends TestRunner {
                 result = createErrorResult(td, i18n.getString("dtr.noResult", url), null);
             }
 
-            if(this.crashOnly && result.getStatus().getType() == Status.FAILED && ! this.didCrash(td)){
-                result = new TestResult(td, new Status(Status.PASSED, "only a regular fail"));
-                if(result.getStatus().getType() == Status.PASSED){
-                }
-            }
-
             if (!scriptUsesNotifier) {
                 try {
                     notifyFinishedTest(result);
@@ -226,27 +214,11 @@ public class DefaultTestRunner extends TestRunner {
             }
         }
 
-        return result.getStatus().getType() == Status.PASSED;
+        return result;
     }
 
-    /**
-     * checks for the occurrence of hs_err_pid file that suggest that crash happened during execution
-     **/
-    private boolean didCrash(TestDescription td){
-        return !getCrashes(td).isEmpty();
-    }
-
-    private List<Crash> getCrashes(TestDescription td){
-        Pattern pattern = Pattern.compile("^hs_err_pid(\\d+)\\.log");
-        List<String> hs_errs = Arrays.stream(td.getDir().list()).filter(pattern.asPredicate()).toList();
-        List<Crash> crashes = new ArrayList<>();
-        for(String hs_err : hs_errs){
-            Matcher m = pattern.matcher(hs_err);
-            m.find();
-            String pid = m.group(1);
-            crashes.add(new Crash(td, hs_err, pid));
-        }
-        return crashes;
+    protected boolean runTest(TestDescription td) {
+        return runTestGiveResult(td).getStatus().getType() == Status.PASSED;
     }
 
     private TestResult createErrorResult(TestDescription td, String reason, Throwable t) { // make more i18n
